Break up large files, add AGENTS.md code standards (#157)
- Add AGENTS.md with code standards (file size, JSX, i18n, TypeScript,
  React/Next.js, pnpm); symlink CLAUDE.md to it
- Split firebase/schema.ts → schema/{lobby,game,player-state,index}.ts
- Split werewolf/actions.ts → actions/{types,helpers,start-night,
  start-day,set-night-phase,set-night-target,confirm-night-target,
  reveal-investigation-result,mark-player,index}.ts
- Split werewolf/actions.test.ts → per-action test files with shared
  fixtures in actions/test-helpers.ts
- All existing import paths unchanged via barrel index.ts exports
