--- app/src/lib/game-modes/werewolf/actions.test.ts
@@ -1,1576 +1,0 @@
-import { describe, it, expect } from "vitest";
-import { GameMode, GameStatus, ShowRolesInPlay } from "@/lib/types";
-import type { Game } from "@/lib/types";
-import { WerewolfPhase } from "./types";
-import type {
-  WerewolfTurnState,
-  WerewolfNighttimePhase,
-  AnyNightAction,
-  TeamNightAction,
-} from "./types";
-import { WerewolfRole } from "./roles";
-import { WerewolfAction, WEREWOLF_ACTIONS } from "./actions";
-import {
-  buildNightPhaseOrder,
-  computeSuggestedTarget,
-  getGroupPhasePlayerIds,
-  resolveNightActions,
-} from "./utils";
-
-function makePlayingGame(
-  turnState: WerewolfTurnState,
-  overrides: Partial<Game> = {},
-): Game {
-  return {
-    id: "game-1",
-    lobbyId: "lobby-1",
-    gameMode: GameMode.Werewolf,
-    status: { type: GameStatus.Playing, turnState },
-    players: [
-      { id: "p1", name: "Alice", sessionId: "s1", visibleRoles: [] },
-      { id: "p2", name: "Bob", sessionId: "s2", visibleRoles: [] },
-      { id: "p3", name: "Charlie", sessionId: "s3", visibleRoles: [] },
-    ],
-    roleAssignments: [
-      { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-      { playerId: "p2", roleDefinitionId: WerewolfRole.Seer },
-      { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-    ],
-    configuredRoleSlots: [],
-    showRolesInPlay: ShowRolesInPlay.None,
-    ownerPlayerId: "owner-1",
-    ...overrides,
-  };
-}
-
-function makeNightState(
-  overrides: Partial<{
-    turn: number;
-    nightActions: Record<string, AnyNightAction>;
-    deadPlayerIds: string[];
-    currentPhaseIndex: number;
-    nightPhaseOrder: string[];
-  }> = {},
-): WerewolfTurnState {
-  return {
-    turn: overrides.turn ?? 1,
-    phase: {
-      type: WerewolfPhase.Nighttime,
-      startedAt: 1000,
-      nightPhaseOrder: overrides.nightPhaseOrder ?? [
-        WerewolfRole.Werewolf,
-        WerewolfRole.Seer,
-      ],
-      currentPhaseIndex: overrides.currentPhaseIndex ?? 0,
-      nightActions: overrides.nightActions ?? {},
-    },
-    deadPlayerIds: overrides.deadPlayerIds ?? [],
-  };
-}
-
-const nightTurnState: WerewolfTurnState = {
-  turn: 1,
-  phase: {
-    type: WerewolfPhase.Nighttime,
-    startedAt: 1000,
-    nightPhaseOrder: [WerewolfRole.Werewolf, WerewolfRole.Seer],
-    currentPhaseIndex: 0,
-    nightActions: {},
-  },
-  deadPlayerIds: [],
-};
-
-const nightTurn2State: WerewolfTurnState = {
-  turn: 2,
-  phase: {
-    type: WerewolfPhase.Nighttime,
-    startedAt: 1000,
-    nightPhaseOrder: [WerewolfRole.Werewolf, WerewolfRole.Seer],
-    currentPhaseIndex: 0,
-    nightActions: {},
-  },
-  deadPlayerIds: [],
-};
-
-const dayTurnState: WerewolfTurnState = {
-  turn: 1,
-  phase: {
-    type: WerewolfPhase.Daytime,
-    startedAt: 1000,
-    nightActions: {},
-  },
-  deadPlayerIds: [],
-};
-
-// ---------------------------------------------------------------------------
-// StartNight
-// ---------------------------------------------------------------------------
-
-describe("WerewolfAction.StartNight", () => {
-  const action = WEREWOLF_ACTIONS[WerewolfAction.StartNight];
-
-  describe("isValid", () => {
-    it("returns true during daytime when called by owner", () => {
-      const game = makePlayingGame(dayTurnState);
-      expect(action.isValid(game, "owner-1", null)).toBe(true);
-    });
-
-    it("returns false during nighttime", () => {
-      const game = makePlayingGame(nightTurnState);
-      expect(action.isValid(game, "owner-1", null)).toBe(false);
-    });
-
-    it("returns false when called by non-owner", () => {
-      const game = makePlayingGame(dayTurnState);
-      expect(action.isValid(game, "player-2", null)).toBe(false);
-    });
-
-    it("returns false when game is not Playing", () => {
-      const game: Game = {
-        id: "game-1",
-        lobbyId: "lobby-1",
-        gameMode: GameMode.Werewolf,
-        status: { type: GameStatus.Starting },
-        players: [],
-        roleAssignments: [],
-        configuredRoleSlots: [],
-        showRolesInPlay: ShowRolesInPlay.None,
-        ownerPlayerId: "owner-1",
-      };
-      expect(action.isValid(game, "owner-1", null)).toBe(false);
-    });
-  });
-
-  describe("apply", () => {
-    it("transitions to nighttime on the next turn", () => {
-      const game = makePlayingGame(dayTurnState);
-      action.apply(game, null, "owner-1");
-      const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-      expect(ts.turn).toBe(2);
-      expect(ts.phase.type).toBe(WerewolfPhase.Nighttime);
-    });
-
-    it("builds nightPhaseOrder with a single Werewolf group phase key", () => {
-      const game = makePlayingGame(dayTurnState);
-      action.apply(game, null, "owner-1");
-      const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-      const phase = ts.phase as WerewolfNighttimePhase;
-      expect(phase.nightPhaseOrder).toContain(WerewolfRole.Werewolf);
-      expect(phase.nightPhaseOrder).toContain(WerewolfRole.Seer);
-      expect(
-        phase.nightPhaseOrder.filter(
-          (k) => k === (WerewolfRole.Werewolf as string),
-        ),
-      ).toHaveLength(1);
-      expect(phase.currentPhaseIndex).toBe(0);
-    });
-
-    it("sets startedAt to a recent timestamp", () => {
-      const before = Date.now();
-      const game = makePlayingGame(dayTurnState);
-      action.apply(game, null, "owner-1");
-      const after = Date.now();
-      const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-      const phase = ts.phase as WerewolfNighttimePhase;
-      expect(phase.startedAt).toBeGreaterThanOrEqual(before);
-      expect(phase.startedAt).toBeLessThanOrEqual(after);
-    });
-  });
-});
-
-// ---------------------------------------------------------------------------
-// StartDay
-// ---------------------------------------------------------------------------
-
-describe("WerewolfAction.StartDay", () => {
-  const action = WEREWOLF_ACTIONS[WerewolfAction.StartDay];
-
-  describe("isValid", () => {
-    it("returns true during nighttime when called by owner", () => {
-      const game = makePlayingGame(nightTurnState);
-      expect(action.isValid(game, "owner-1", null)).toBe(true);
-    });
-
-    it("returns false during daytime", () => {
-      const game = makePlayingGame(dayTurnState);
-      expect(action.isValid(game, "owner-1", null)).toBe(false);
-    });
-
-    it("returns false when called by non-owner", () => {
-      const game = makePlayingGame(nightTurnState);
-      expect(action.isValid(game, "player-2", null)).toBe(false);
-    });
-  });
-
-  describe("apply", () => {
-    it("transitions to daytime on the same turn", () => {
-      const game = makePlayingGame(nightTurnState);
-      action.apply(game, null, "owner-1");
-      const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-      expect(ts.turn).toBe(1);
-      expect(ts.phase.type).toBe(WerewolfPhase.Daytime);
-    });
-
-    it("sets startedAt to a recent timestamp", () => {
-      const before = Date.now();
-      const game = makePlayingGame(nightTurnState);
-      action.apply(game, null, "owner-1");
-      const after = Date.now();
-      const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-      const phase = ts.phase as { startedAt: number };
-      expect(phase.startedAt).toBeGreaterThanOrEqual(before);
-      expect(phase.startedAt).toBeLessThanOrEqual(after);
-    });
-
-    it("resolves night actions and adds killed players to deadPlayerIds", () => {
-      const game = makePlayingGame(
-        makeNightState({
-          nightActions: {
-            [WerewolfRole.Werewolf]: {
-              votes: [],
-              suggestedTargetId: "p2",
-            },
-          },
-        }),
-      );
-      action.apply(game, null, "owner-1");
-      const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-      expect(ts.deadPlayerIds).toContain("p2");
-    });
-
-    it("does not kill a player protected by Bodyguard", () => {
-      const game = makePlayingGame(
-        makeNightState({
-          nightActions: {
-            [WerewolfRole.Werewolf]: {
-              votes: [],
-              suggestedTargetId: "p2",
-            },
-            [WerewolfRole.Bodyguard]: { targetPlayerId: "p2" },
-          },
-        }),
-        {
-          roleAssignments: [
-            { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-            { playerId: "p2", roleDefinitionId: WerewolfRole.Seer },
-            { playerId: "p3", roleDefinitionId: WerewolfRole.Bodyguard },
-          ],
-        },
-      );
-      action.apply(game, null, "owner-1");
-      const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-      expect(ts.deadPlayerIds).not.toContain("p2");
-    });
-
-    it("stores nightResolution in the daytime phase", () => {
-      const game = makePlayingGame(
-        makeNightState({
-          nightActions: {
-            [WerewolfRole.Werewolf]: {
-              votes: [],
-              suggestedTargetId: "p2",
-            },
-          },
-        }),
-      );
-      action.apply(game, null, "owner-1");
-      const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-      const phase = ts.phase as import("./types").WerewolfDaytimePhase;
-      expect(phase.nightResolution).toBeDefined();
-      expect(phase.nightResolution).toHaveLength(1);
-      expect(phase.nightResolution![0]).toMatchObject({
-        type: "killed",
-        targetPlayerId: "p2",
-        died: true,
-      });
-    });
-
-    it("omits nightResolution when there are no night actions", () => {
-      const game = makePlayingGame(nightTurnState);
-      action.apply(game, null, "owner-1");
-      const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-      const phase = ts.phase as import("./types").WerewolfDaytimePhase;
-      expect(phase.nightResolution).toBeUndefined();
-    });
-  });
-});
-
-// ---------------------------------------------------------------------------
-// SetNightPhase
-// ---------------------------------------------------------------------------
-
-describe("WerewolfAction.SetNightPhase", () => {
-  const action = WEREWOLF_ACTIONS[WerewolfAction.SetNightPhase];
-
-  describe("isValid", () => {
-    it("returns true for a valid index during nighttime", () => {
-      const game = makePlayingGame(nightTurnState);
-      expect(action.isValid(game, "owner-1", { phaseIndex: 1 })).toBe(true);
-    });
-
-    it("returns true for index 0", () => {
-      const game = makePlayingGame(nightTurnState);
-      expect(action.isValid(game, "owner-1", { phaseIndex: 0 })).toBe(true);
-    });
-
-    it("returns false for an out-of-bounds index", () => {
-      const game = makePlayingGame(nightTurnState);
-      expect(action.isValid(game, "owner-1", { phaseIndex: 2 })).toBe(false);
-    });
-
-    it("returns false for a negative index", () => {
-      const game = makePlayingGame(nightTurnState);
-      expect(action.isValid(game, "owner-1", { phaseIndex: -1 })).toBe(false);
-    });
-
-    it("returns false during daytime", () => {
-      const game = makePlayingGame(dayTurnState);
-      expect(action.isValid(game, "owner-1", { phaseIndex: 0 })).toBe(false);
-    });
-
-    it("returns false when called by non-owner", () => {
-      const game = makePlayingGame(nightTurnState);
-      expect(action.isValid(game, "player-2", { phaseIndex: 0 })).toBe(false);
-    });
-
-    it("returns false when phaseIndex is not a number", () => {
-      const game = makePlayingGame(nightTurnState);
-      expect(action.isValid(game, "owner-1", { phaseIndex: "0" })).toBe(false);
-    });
-  });
-
-  describe("apply", () => {
-    it("updates currentPhaseIndex", () => {
-      const game = makePlayingGame(nightTurnState);
-      action.apply(game, { phaseIndex: 1 }, "owner-1");
-      const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-      const phase = ts.phase as WerewolfNighttimePhase;
-      expect(phase.currentPhaseIndex).toBe(1);
-    });
-
-    it("preserves nightPhaseOrder and turn", () => {
-      const game = makePlayingGame(nightTurnState);
-      action.apply(game, { phaseIndex: 1 }, "owner-1");
-      const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-      const phase = ts.phase as WerewolfNighttimePhase;
-      expect(ts.turn).toBe(1);
-      expect(phase.nightPhaseOrder).toEqual([
-        WerewolfRole.Werewolf,
-        WerewolfRole.Seer,
-      ]);
-    });
-
-    it("resets startedAt to a recent timestamp", () => {
-      const before = Date.now();
-      const game = makePlayingGame(nightTurnState);
-      action.apply(game, { phaseIndex: 1 }, "owner-1");
-      const after = Date.now();
-      const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-      const phase = ts.phase as WerewolfNighttimePhase;
-      expect(phase.startedAt).toBeGreaterThanOrEqual(before);
-      expect(phase.startedAt).toBeLessThanOrEqual(after);
-    });
-  });
-});
-
-// ---------------------------------------------------------------------------
-// SetNightTarget — solo roles (existing behavior with hardcoded phase keys)
-// ---------------------------------------------------------------------------
-
-describe("WerewolfAction.SetNightTarget", () => {
-  const action = WEREWOLF_ACTIONS[WerewolfAction.SetNightTarget];
-
-  describe("isValid — owner", () => {
-    it("allows owner to set a target with explicit roleId on turn 2+", () => {
-      const game = makePlayingGame(nightTurn2State);
-      expect(
-        action.isValid(game, "owner-1", {
-          roleId: WerewolfRole.Seer,
-          targetPlayerId: "p1",
-        }),
-      ).toBe(true);
-    });
-  });
-
-  describe("isValid — player (solo role)", () => {
-    it("returns true when active player targets a valid player on turn 2+", () => {
-      const game = makePlayingGame(nightTurn2State);
-      expect(action.isValid(game, "p1", { targetPlayerId: "p2" })).toBe(true);
-    });
-
-    it("returns true when active player clears target (undefined targetPlayerId)", () => {
-      const game = makePlayingGame(nightTurn2State);
-      expect(action.isValid(game, "p1", { targetPlayerId: undefined })).toBe(
-        true,
-      );
-    });
-
-    it("returns false on first turn", () => {
-      const game = makePlayingGame(nightTurnState);
-      expect(action.isValid(game, "p1", { targetPlayerId: "p2" })).toBe(false);
-    });
-
-    it("returns false when caller is not the active role", () => {
-      const game = makePlayingGame(nightTurn2State);
-      expect(action.isValid(game, "p2", { targetPlayerId: "p1" })).toBe(false);
-    });
-
-    it("returns false when targeting the game owner", () => {
-      const game = makePlayingGame(nightTurn2State);
-      expect(action.isValid(game, "p1", { targetPlayerId: "owner-1" })).toBe(
-        false,
-      );
-    });
-
-    it("returns false when targeting a non-existent player", () => {
-      const game = makePlayingGame(nightTurn2State);
-      expect(action.isValid(game, "p1", { targetPlayerId: "unknown" })).toBe(
-        false,
-      );
-    });
-
-    it("returns false when targeting a dead player", () => {
-      const game = makePlayingGame(
-        makeNightState({ turn: 2, deadPlayerIds: ["p2"] }),
-      );
-      expect(action.isValid(game, "p1", { targetPlayerId: "p2" })).toBe(false);
-    });
-
-    it("returns false during daytime", () => {
-      const game = makePlayingGame(dayTurnState);
-      expect(action.isValid(game, "p1", { targetPlayerId: "p2" })).toBe(false);
-    });
-
-    it("returns false when targetPlayerId is not a string", () => {
-      const game = makePlayingGame(nightTurn2State);
-      expect(action.isValid(game, "p1", { targetPlayerId: 123 })).toBe(false);
-    });
-
-    it("returns false when caller has no role assignment", () => {
-      const game = makePlayingGame(nightTurn2State, { roleAssignments: [] });
-      expect(action.isValid(game, "p1", { targetPlayerId: "p2" })).toBe(false);
-    });
-
-    it("returns false when player's target is already confirmed", () => {
-      const game = makePlayingGame(
-        makeNightState({
-          turn: 2,
-          nightActions: {
-            [WerewolfRole.Werewolf]: {
-              targetPlayerId: "p2",
-              confirmed: true,
-            },
-          },
-        }),
-      );
-      expect(action.isValid(game, "p1", { targetPlayerId: "p3" })).toBe(false);
-    });
-
-    it("allows owner to override a confirmed target", () => {
-      const game = makePlayingGame(
-        makeNightState({
-          turn: 2,
-          nightActions: {
-            [WerewolfRole.Werewolf]: {
-              targetPlayerId: "p2",
-              confirmed: true,
-            },
-          },
-        }),
-      );
-      expect(
-        action.isValid(game, "owner-1", {
-          roleId: WerewolfRole.Werewolf,
-          targetPlayerId: "p3",
-        }),
-      ).toBe(true);
-    });
-  });
-
-  describe("apply — owner (explicit roleId, solo)", () => {
-    it("sets the night action for the specified role", () => {
-      const game = makePlayingGame(nightTurn2State);
-      action.apply(
-        game,
-        { roleId: WerewolfRole.Seer, targetPlayerId: "p1" },
-        "owner-1",
-      );
-      const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-      const phase = ts.phase as WerewolfNighttimePhase;
-      expect(phase.nightActions[WerewolfRole.Seer]).toEqual({
-        targetPlayerId: "p1",
-      });
-    });
-
-    it("clears the night action when targetPlayerId is undefined", () => {
-      const game = makePlayingGame(
-        makeNightState({
-          turn: 2,
-          nightActions: {
-            [WerewolfRole.Seer]: { targetPlayerId: "p1" },
-          },
-        }),
-      );
-      action.apply(
-        game,
-        { roleId: WerewolfRole.Seer, targetPlayerId: undefined },
-        "owner-1",
-      );
-      const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-      const phase = ts.phase as WerewolfNighttimePhase;
-      expect(phase.nightActions[WerewolfRole.Seer]).toBeUndefined();
-    });
-
-    it("does not affect other roles when clearing", () => {
-      const game = makePlayingGame(
-        makeNightState({
-          turn: 2,
-          nightActions: {
-            [WerewolfRole.Seer]: { targetPlayerId: "p1" },
-            [WerewolfRole.Bodyguard]: { targetPlayerId: "p3" },
-          },
-        }),
-      );
-      action.apply(
-        game,
-        { roleId: WerewolfRole.Seer, targetPlayerId: undefined },
-        "owner-1",
-      );
-      const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-      const phase = ts.phase as WerewolfNighttimePhase;
-      expect(phase.nightActions[WerewolfRole.Seer]).toBeUndefined();
-      expect(phase.nightActions[WerewolfRole.Bodyguard]).toEqual({
-        targetPlayerId: "p3",
-      });
-    });
-  });
-
-  describe("apply — player (inferred roleId, solo)", () => {
-    // Use Seer (p2) as the active solo role by putting Seer first in the phase order.
-    const seerActiveState = makeNightState({
-      turn: 2,
-      nightPhaseOrder: [WerewolfRole.Seer, WerewolfRole.Werewolf],
-      currentPhaseIndex: 0,
-    });
-
-    it("sets the night action for the active role", () => {
-      const game = makePlayingGame(seerActiveState);
-      action.apply(game, { targetPlayerId: "p1" }, "p2");
-      const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-      const phase = ts.phase as WerewolfNighttimePhase;
-      expect(phase.nightActions[WerewolfRole.Seer]).toEqual({
-        targetPlayerId: "p1",
-      });
-    });
-
-    it("clears the night action when targetPlayerId is undefined", () => {
-      const game = makePlayingGame(
-        makeNightState({
-          turn: 2,
-          nightPhaseOrder: [WerewolfRole.Seer, WerewolfRole.Werewolf],
-          currentPhaseIndex: 0,
-          nightActions: {
-            [WerewolfRole.Seer]: { targetPlayerId: "p1" },
-          },
-        }),
-      );
-      action.apply(game, { targetPlayerId: undefined }, "p2");
-      const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-      const phase = ts.phase as WerewolfNighttimePhase;
-      expect(phase.nightActions[WerewolfRole.Seer]).toBeUndefined();
-    });
-
-    it("overwrites a previous target", () => {
-      const game = makePlayingGame(
-        makeNightState({
-          turn: 2,
-          nightPhaseOrder: [WerewolfRole.Seer, WerewolfRole.Werewolf],
-          currentPhaseIndex: 0,
-          nightActions: {
-            [WerewolfRole.Seer]: { targetPlayerId: "p3" },
-          },
-        }),
-      );
-      action.apply(game, { targetPlayerId: "p1" }, "p2");
-      const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-      const phase = ts.phase as WerewolfNighttimePhase;
-      expect(phase.nightActions[WerewolfRole.Seer]).toEqual({
-        targetPlayerId: "p1",
-      });
-    });
-  });
-});
-
-// ---------------------------------------------------------------------------
-// ConfirmNightTarget — solo roles
-// ---------------------------------------------------------------------------
-
-describe("WerewolfAction.ConfirmNightTarget", () => {
-  const action = WEREWOLF_ACTIONS[WerewolfAction.ConfirmNightTarget];
-
-  describe("isValid (solo)", () => {
-    // Use Seer (p2) as the active solo role.
-    const seerActiveState = makeNightState({
-      turn: 2,
-      nightPhaseOrder: [WerewolfRole.Seer, WerewolfRole.Werewolf],
-      currentPhaseIndex: 0,
-    });
-
-    it("returns true when active player has an unconfirmed target on turn 2+", () => {
-      const game = makePlayingGame(
-        makeNightState({
-          turn: 2,
-          nightPhaseOrder: [WerewolfRole.Seer, WerewolfRole.Werewolf],
-          currentPhaseIndex: 0,
-          nightActions: {
-            [WerewolfRole.Seer]: { targetPlayerId: "p1" },
-          },
-        }),
-      );
-      expect(action.isValid(game, "p2", null)).toBe(true);
-    });
-
-    it("returns false when no target is set", () => {
-      const game = makePlayingGame(seerActiveState);
-      expect(action.isValid(game, "p2", null)).toBe(false);
-    });
-
-    it("returns false when target is already confirmed", () => {
-      const game = makePlayingGame(
-        makeNightState({
-          turn: 2,
-          nightPhaseOrder: [WerewolfRole.Seer, WerewolfRole.Werewolf],
-          currentPhaseIndex: 0,
-          nightActions: {
-            [WerewolfRole.Seer]: { targetPlayerId: "p1", confirmed: true },
-          },
-        }),
-      );
-      expect(action.isValid(game, "p2", null)).toBe(false);
-    });
-
-    it("returns false on first turn", () => {
-      const game = makePlayingGame(
-        makeNightState({
-          turn: 1,
-          nightPhaseOrder: [WerewolfRole.Seer, WerewolfRole.Werewolf],
-          currentPhaseIndex: 0,
-          nightActions: {
-            [WerewolfRole.Seer]: { targetPlayerId: "p1" },
-          },
-        }),
-      );
-      expect(action.isValid(game, "p2", null)).toBe(false);
-    });
-
-    it("returns false when caller is not the active role", () => {
-      const game = makePlayingGame(
-        makeNightState({
-          turn: 2,
-          nightPhaseOrder: [WerewolfRole.Seer, WerewolfRole.Werewolf],
-          currentPhaseIndex: 0,
-          nightActions: {
-            [WerewolfRole.Seer]: { targetPlayerId: "p1" },
-          },
-        }),
-      );
-      expect(action.isValid(game, "p1", null)).toBe(false);
-    });
-
-    it("returns false when caller has no role assignment", () => {
-      const game = makePlayingGame(
-        makeNightState({
-          turn: 2,
-          nightPhaseOrder: [WerewolfRole.Seer, WerewolfRole.Werewolf],
-          currentPhaseIndex: 0,
-          nightActions: {
-            [WerewolfRole.Seer]: { targetPlayerId: "p1" },
-          },
-        }),
-        { roleAssignments: [] },
-      );
-      expect(action.isValid(game, "p2", null)).toBe(false);
-    });
-
-    it("returns false during daytime", () => {
-      const game = makePlayingGame(dayTurnState);
-      expect(action.isValid(game, "p2", null)).toBe(false);
-    });
-  });
-
-  describe("apply (solo)", () => {
-    it("sets confirmed to true on the active role's night action", () => {
-      const game = makePlayingGame(
-        makeNightState({
-          turn: 2,
-          nightActions: {
-            [WerewolfRole.Werewolf]: { targetPlayerId: "p2" },
-          },
-        }),
-      );
-      action.apply(game, null, "p1");
-      const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-      const phase = ts.phase as WerewolfNighttimePhase;
-      expect(phase.nightActions[WerewolfRole.Werewolf]).toEqual({
-        targetPlayerId: "p2",
-        confirmed: true,
-      });
-    });
-
-    it("does not affect other roles", () => {
-      const game = makePlayingGame(
-        makeNightState({
-          turn: 2,
-          nightActions: {
-            [WerewolfRole.Werewolf]: { targetPlayerId: "p2" },
-            [WerewolfRole.Seer]: { targetPlayerId: "p3" },
-          },
-        }),
-      );
-      action.apply(game, null, "p1");
-      const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-      const phase = ts.phase as WerewolfNighttimePhase;
-      expect(phase.nightActions[WerewolfRole.Seer]).toEqual({
-        targetPlayerId: "p3",
-      });
-    });
-  });
-});
-
-// ---------------------------------------------------------------------------
-// Team targeting — buildNightPhaseOrder, SetNightTarget, ConfirmNightTarget
-// ---------------------------------------------------------------------------
-
-const TEAM_BAD_KEY = WerewolfRole.Werewolf;
-
-function makeTeamGame(
-  turnState: WerewolfTurnState,
-  overrides: Partial<Game> = {},
-): Game {
-  return {
-    id: "game-1",
-    lobbyId: "lobby-1",
-    gameMode: GameMode.Werewolf,
-    status: { type: GameStatus.Playing, turnState },
-    players: [
-      {
-        id: "w1",
-        name: "Wolf1",
-        sessionId: "sw1",
-        visibleRoles: [
-          { playerId: "w2", roleDefinitionId: WerewolfRole.Werewolf },
-        ],
-      },
-      {
-        id: "w2",
-        name: "Wolf2",
-        sessionId: "sw2",
-        visibleRoles: [
-          { playerId: "w1", roleDefinitionId: WerewolfRole.Werewolf },
-        ],
-      },
-      { id: "p3", name: "Seer", sessionId: "s3", visibleRoles: [] },
-      { id: "p4", name: "Villager", sessionId: "s4", visibleRoles: [] },
-    ],
-    roleAssignments: [
-      { playerId: "w1", roleDefinitionId: WerewolfRole.Werewolf },
-      { playerId: "w2", roleDefinitionId: WerewolfRole.Werewolf },
-      { playerId: "p3", roleDefinitionId: WerewolfRole.Seer },
-      { playerId: "p4", roleDefinitionId: WerewolfRole.Villager },
-    ],
-    configuredRoleSlots: [],
-    showRolesInPlay: ShowRolesInPlay.None,
-    ownerPlayerId: "owner-1",
-    ...overrides,
-  };
-}
-
-function makeTeamNightState(
-  overrides: Partial<{
-    turn: number;
-    nightActions: Record<string, AnyNightAction>;
-    deadPlayerIds: string[];
-    currentPhaseIndex: number;
-  }> = {},
-): WerewolfTurnState {
-  return {
-    turn: overrides.turn ?? 2,
-    phase: {
-      type: WerewolfPhase.Nighttime,
-      startedAt: 1000,
-      nightPhaseOrder: [TEAM_BAD_KEY, WerewolfRole.Seer],
-      currentPhaseIndex: overrides.currentPhaseIndex ?? 0,
-      nightActions: overrides.nightActions ?? {},
-    },
-    deadPlayerIds: overrides.deadPlayerIds ?? [],
-  };
-}
-
-describe("buildNightPhaseOrder — team targeting", () => {
-  it("emits a single Werewolf group phase key for multiple werewolves", () => {
-    const assignments = [
-      { playerId: "w1", roleDefinitionId: WerewolfRole.Werewolf },
-      { playerId: "w2", roleDefinitionId: WerewolfRole.Werewolf },
-      { playerId: "p3", roleDefinitionId: WerewolfRole.Seer },
-    ];
-    const order = buildNightPhaseOrder(2, assignments);
-    expect(order).toContain(WerewolfRole.Werewolf);
-    expect(order).toContain(WerewolfRole.Seer);
-    expect(
-      order.filter((k) => k === (WerewolfRole.Werewolf as string)),
-    ).toHaveLength(1);
-  });
-
-  it("emits group phase key even for a single werewolf", () => {
-    const assignments = [
-      { playerId: "w1", roleDefinitionId: WerewolfRole.Werewolf },
-      { playerId: "p3", roleDefinitionId: WerewolfRole.Seer },
-    ];
-    const order = buildNightPhaseOrder(2, assignments);
-    expect(order).toContain(WerewolfRole.Werewolf);
-    expect(
-      order.filter((k) => k === (WerewolfRole.Werewolf as string)),
-    ).toHaveLength(1);
-  });
-
-  it("does not group Chupacabra into team phase", () => {
-    const assignments = [
-      { playerId: "w1", roleDefinitionId: WerewolfRole.Werewolf },
-      { playerId: "c1", roleDefinitionId: WerewolfRole.Chupacabra },
-      { playerId: "p3", roleDefinitionId: WerewolfRole.Seer },
-    ];
-    const order = buildNightPhaseOrder(2, assignments);
-    expect(order).toContain(TEAM_BAD_KEY);
-    expect(order).toContain(WerewolfRole.Chupacabra);
-  });
-});
-
-describe("computeSuggestedTarget", () => {
-  it("returns the most-voted target", () => {
-    expect(
-      computeSuggestedTarget([
-        { playerId: "w1", targetPlayerId: "p3" },
-        { playerId: "w2", targetPlayerId: "p3" },
-      ]),
-    ).toBe("p3");
-  });
-
-  it("returns undefined on a tie", () => {
-    expect(
-      computeSuggestedTarget([
-        { playerId: "w1", targetPlayerId: "p3" },
-        { playerId: "w2", targetPlayerId: "p4" },
-      ]),
-    ).toBeUndefined();
-  });
-
-  it("returns undefined for no votes", () => {
-    expect(computeSuggestedTarget([])).toBeUndefined();
-  });
-
-  it("returns the single vote when only one exists", () => {
-    expect(
-      computeSuggestedTarget([{ playerId: "w1", targetPlayerId: "p4" }]),
-    ).toBe("p4");
-  });
-});
-
-describe("getGroupPhasePlayerIds", () => {
-  const assignments = [
-    { playerId: "w1", roleDefinitionId: WerewolfRole.Werewolf },
-    { playerId: "w2", roleDefinitionId: WerewolfRole.Werewolf },
-    { playerId: "p3", roleDefinitionId: WerewolfRole.Seer },
-  ];
-
-  it("returns alive team players with teamTargeting", () => {
-    expect(
-      getGroupPhasePlayerIds(assignments, WerewolfRole.Werewolf, []),
-    ).toEqual(["w1", "w2"]);
-  });
-
-  it("excludes dead players", () => {
-    expect(
-      getGroupPhasePlayerIds(assignments, WerewolfRole.Werewolf, ["w2"]),
-    ).toEqual(["w1"]);
-  });
-});
-
-describe("SetNightTarget — team phase", () => {
-  const action = WEREWOLF_ACTIONS[WerewolfAction.SetNightTarget];
-
-  it("player can vote in a team phase", () => {
-    const game = makeTeamGame(makeTeamNightState());
-    expect(action.isValid(game, "w1", { targetPlayerId: "p3" })).toBe(true);
-  });
-
-  it("werewolf cannot target another werewolf", () => {
-    const game = makeTeamGame(makeTeamNightState());
-    expect(action.isValid(game, "w1", { targetPlayerId: "w2" })).toBe(false);
-  });
-
-  it("non-team player cannot vote in team phase", () => {
-    const game = makeTeamGame(makeTeamNightState());
-    // p3 is Seer, not on Bad team
-    expect(action.isValid(game, "p3", { targetPlayerId: "p4" })).toBe(false);
-  });
-
-  it("apply creates a TeamNightAction with the voter's entry", () => {
-    const game = makeTeamGame(makeTeamNightState());
-    action.apply(game, { targetPlayerId: "p3" }, "w1");
-    const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-    const phase = ts.phase as WerewolfNighttimePhase;
-    const groupAction = phase.nightActions[TEAM_BAD_KEY] as TeamNightAction;
-    expect(groupAction.votes).toEqual([
-      { playerId: "w1", targetPlayerId: "p3" },
-    ]);
-    expect(groupAction.suggestedTargetId).toBe("p3");
-  });
-
-  it("second voter updates the TeamNightAction", () => {
-    const game = makeTeamGame(makeTeamNightState());
-    action.apply(game, { targetPlayerId: "p3" }, "w1");
-    action.apply(game, { targetPlayerId: "p4" }, "w2");
-    const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-    const phase = ts.phase as WerewolfNighttimePhase;
-    const groupAction = phase.nightActions[TEAM_BAD_KEY] as TeamNightAction;
-    expect(groupAction.votes).toHaveLength(2);
-    // Tied — no suggested target.
-    expect(groupAction.suggestedTargetId).toBeUndefined();
-  });
-
-  it("voter can change their vote", () => {
-    const game = makeTeamGame(makeTeamNightState());
-    action.apply(game, { targetPlayerId: "p3" }, "w1");
-    action.apply(game, { targetPlayerId: "p4" }, "w1");
-    const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-    const phase = ts.phase as WerewolfNighttimePhase;
-    const groupAction = phase.nightActions[TEAM_BAD_KEY] as TeamNightAction;
-    expect(groupAction.votes).toEqual([
-      { playerId: "w1", targetPlayerId: "p4" },
-    ]);
-  });
-
-  it("voter can clear their vote", () => {
-    const game = makeTeamGame(makeTeamNightState());
-    action.apply(game, { targetPlayerId: "p3" }, "w1");
-    action.apply(game, { targetPlayerId: undefined }, "w1");
-    const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-    const phase = ts.phase as WerewolfNighttimePhase;
-    const groupAction = phase.nightActions[TEAM_BAD_KEY] as TeamNightAction;
-    expect(groupAction.votes).toEqual([]);
-  });
-
-  it("owner override sets all alive team members' votes", () => {
-    const game = makeTeamGame(makeTeamNightState());
-    action.apply(
-      game,
-      { roleId: TEAM_BAD_KEY, targetPlayerId: "p4" },
-      "owner-1",
-    );
-    const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-    const phase = ts.phase as WerewolfNighttimePhase;
-    const groupAction = phase.nightActions[TEAM_BAD_KEY] as TeamNightAction;
-    expect(groupAction.votes).toEqual([
-      { playerId: "w1", targetPlayerId: "p4" },
-      { playerId: "w2", targetPlayerId: "p4" },
-    ]);
-    expect(groupAction.suggestedTargetId).toBe("p4");
-  });
-
-  it("blocks voting after team confirmed", () => {
-    const game = makeTeamGame(
-      makeTeamNightState({
-        nightActions: {
-          [TEAM_BAD_KEY]: {
-            votes: [
-              { playerId: "w1", targetPlayerId: "p3" },
-              { playerId: "w2", targetPlayerId: "p3" },
-            ],
-            suggestedTargetId: "p3",
-            confirmed: true,
-          },
-        },
-      }),
-    );
-    expect(action.isValid(game, "w1", { targetPlayerId: "p4" })).toBe(false);
-  });
-});
-
-describe("ConfirmNightTarget — team phase", () => {
-  const action = WEREWOLF_ACTIONS[WerewolfAction.ConfirmNightTarget];
-
-  it("returns true when all alive team members agree", () => {
-    const game = makeTeamGame(
-      makeTeamNightState({
-        nightActions: {
-          [TEAM_BAD_KEY]: {
-            votes: [
-              { playerId: "w1", targetPlayerId: "p3" },
-              { playerId: "w2", targetPlayerId: "p3" },
-            ],
-            suggestedTargetId: "p3",
-          },
-        },
-      }),
-    );
-    expect(action.isValid(game, "w1", null)).toBe(true);
-  });
-
-  it("returns false when votes disagree", () => {
-    const game = makeTeamGame(
-      makeTeamNightState({
-        nightActions: {
-          [TEAM_BAD_KEY]: {
-            votes: [
-              { playerId: "w1", targetPlayerId: "p3" },
-              { playerId: "w2", targetPlayerId: "p4" },
-            ],
-            suggestedTargetId: undefined,
-          },
-        },
-      }),
-    );
-    expect(action.isValid(game, "w1", null)).toBe(false);
-  });
-
-  it("returns false when not all team members have voted", () => {
-    const game = makeTeamGame(
-      makeTeamNightState({
-        nightActions: {
-          [TEAM_BAD_KEY]: {
-            votes: [{ playerId: "w1", targetPlayerId: "p3" }],
-            suggestedTargetId: "p3",
-          },
-        },
-      }),
-    );
-    expect(action.isValid(game, "w1", null)).toBe(false);
-  });
-
-  it("considers only alive members (dead member excluded)", () => {
-    const game = makeTeamGame(
-      makeTeamNightState({
-        deadPlayerIds: ["w2"],
-        nightActions: {
-          [TEAM_BAD_KEY]: {
-            votes: [{ playerId: "w1", targetPlayerId: "p3" }],
-            suggestedTargetId: "p3",
-          },
-        },
-      }),
-    );
-    expect(action.isValid(game, "w1", null)).toBe(true);
-  });
-
-  it("returns false when already confirmed", () => {
-    const game = makeTeamGame(
-      makeTeamNightState({
-        nightActions: {
-          [TEAM_BAD_KEY]: {
-            votes: [
-              { playerId: "w1", targetPlayerId: "p3" },
-              { playerId: "w2", targetPlayerId: "p3" },
-            ],
-            suggestedTargetId: "p3",
-            confirmed: true,
-          },
-        },
-      }),
-    );
-    expect(action.isValid(game, "w1", null)).toBe(false);
-  });
-
-  it("apply sets confirmed to true on the team action", () => {
-    const game = makeTeamGame(
-      makeTeamNightState({
-        nightActions: {
-          [TEAM_BAD_KEY]: {
-            votes: [
-              { playerId: "w1", targetPlayerId: "p3" },
-              { playerId: "w2", targetPlayerId: "p3" },
-            ],
-            suggestedTargetId: "p3",
-          },
-        },
-      }),
-    );
-    action.apply(game, null, "w1");
-    const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-    const phase = ts.phase as WerewolfNighttimePhase;
-    const groupAction = phase.nightActions[TEAM_BAD_KEY] as TeamNightAction;
-    expect(groupAction.confirmed).toBe(true);
-  });
-});
-
-// ---------------------------------------------------------------------------
-// Bodyguard self-targeting
-// ---------------------------------------------------------------------------
-
-describe("SetNightTarget — Bodyguard self-targeting", () => {
-  const action = WEREWOLF_ACTIONS[WerewolfAction.SetNightTarget];
-
-  function makeBodyguardGame(nightState: WerewolfTurnState): Game {
-    return makePlayingGame(nightState, {
-      players: [
-        { id: "p1", name: "Alice", sessionId: "s1", visibleRoles: [] },
-        { id: "p2", name: "Bob", sessionId: "s2", visibleRoles: [] },
-        { id: "p3", name: "Charlie", sessionId: "s3", visibleRoles: [] },
-      ],
-      roleAssignments: [
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Bodyguard },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Seer },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-      ],
-    });
-  }
-
-  const bodyguardNightState = makeNightState({
-    turn: 2,
-    nightPhaseOrder: [WerewolfRole.Bodyguard],
-  });
-
-  it("Bodyguard can target themselves", () => {
-    const game = makeBodyguardGame(bodyguardNightState);
-    expect(action.isValid(game, "p1", { targetPlayerId: "p1" })).toBe(true);
-  });
-
-  it("Attack role cannot target themselves", () => {
-    const game = makePlayingGame(nightTurn2State);
-    // p1 is Werewolf (Attack), targeting themselves
-    expect(action.isValid(game, "p1", { targetPlayerId: "p1" })).toBe(false);
-  });
-
-  it("Bodyguard self-protection is applied by resolveNightActions", () => {
-    const assignments = [
-      { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-      { playerId: "p2", roleDefinitionId: WerewolfRole.Bodyguard },
-    ];
-    const nightActions = {
-      [WerewolfRole.Werewolf]: { votes: [], suggestedTargetId: "p2" },
-      [WerewolfRole.Bodyguard]: { targetPlayerId: "p2" },
-    };
-    const events = resolveNightActions(nightActions, assignments, []);
-    const event = events.find((e) => e.targetPlayerId === "p2");
-    expect(event?.type === "killed" && event.died).toBe(false);
-  });
-});
-
-// ---------------------------------------------------------------------------
-// RevealInvestigationResult
-// ---------------------------------------------------------------------------
-
-describe("RevealInvestigationResult", () => {
-  const action = WEREWOLF_ACTIONS[WerewolfAction.RevealInvestigationResult];
-
-  const seerNightPhaseOrder = [WerewolfRole.Seer];
-
-  function makeSeerGame(nightState: WerewolfTurnState): Game {
-    return makePlayingGame(nightState, {
-      players: [
-        { id: "p1", name: "Alice", sessionId: "s1", visibleRoles: [] },
-        { id: "p2", name: "Bob", sessionId: "s2", visibleRoles: [] },
-        { id: "p3", name: "Charlie", sessionId: "s3", visibleRoles: [] },
-      ],
-      roleAssignments: [
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Seer },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-      ],
-    });
-  }
-
-  it("is valid when Seer has confirmed an investigation", () => {
-    const game = makeSeerGame(
-      makeNightState({
-        turn: 2,
-        nightPhaseOrder: seerNightPhaseOrder,
-        nightActions: {
-          [WerewolfRole.Seer]: { targetPlayerId: "p2", confirmed: true },
-        },
-      }),
-    );
-    expect(action.isValid(game, "owner-1", null)).toBe(true);
-  });
-
-  it("is invalid when action is not confirmed", () => {
-    const game = makeSeerGame(
-      makeNightState({
-        turn: 2,
-        nightPhaseOrder: seerNightPhaseOrder,
-        nightActions: {
-          [WerewolfRole.Seer]: { targetPlayerId: "p2" },
-        },
-      }),
-    );
-    expect(action.isValid(game, "owner-1", null)).toBe(false);
-  });
-
-  it("is invalid when result is already revealed", () => {
-    const game = makeSeerGame(
-      makeNightState({
-        turn: 2,
-        nightPhaseOrder: seerNightPhaseOrder,
-        nightActions: {
-          [WerewolfRole.Seer]: {
-            targetPlayerId: "p2",
-            confirmed: true,
-            resultRevealed: true,
-          },
-        },
-      }),
-    );
-    expect(action.isValid(game, "owner-1", null)).toBe(false);
-  });
-
-  it("is invalid when active phase is not an Investigate role", () => {
-    const game = makePlayingGame(
-      makeNightState({
-        turn: 2,
-        nightPhaseOrder: [WerewolfRole.Bodyguard],
-        nightActions: {
-          [WerewolfRole.Bodyguard]: { targetPlayerId: "p3", confirmed: true },
-        },
-      }),
-      {
-        roleAssignments: [
-          { playerId: "p1", roleDefinitionId: WerewolfRole.Werewolf },
-          { playerId: "p2", roleDefinitionId: WerewolfRole.Seer },
-          { playerId: "p3", roleDefinitionId: WerewolfRole.Bodyguard },
-        ],
-      },
-    );
-    expect(action.isValid(game, "owner-1", null)).toBe(false);
-  });
-
-  it("is invalid for non-owner callers", () => {
-    const game = makeSeerGame(
-      makeNightState({
-        turn: 2,
-        nightPhaseOrder: seerNightPhaseOrder,
-        nightActions: {
-          [WerewolfRole.Seer]: { targetPlayerId: "p2", confirmed: true },
-        },
-      }),
-    );
-    expect(action.isValid(game, "p1", null)).toBe(false);
-  });
-
-  it("apply sets resultRevealed to true", () => {
-    const game = makeSeerGame(
-      makeNightState({
-        turn: 2,
-        nightPhaseOrder: seerNightPhaseOrder,
-        nightActions: {
-          [WerewolfRole.Seer]: { targetPlayerId: "p2", confirmed: true },
-        },
-      }),
-    );
-    action.apply(game, null, "owner-1");
-    const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-    const phase = ts.phase as WerewolfNighttimePhase;
-    expect(phase.nightActions[WerewolfRole.Seer]).toMatchObject({
-      targetPlayerId: "p2",
-      confirmed: true,
-      resultRevealed: true,
-    });
-  });
-});
-
-// ---------------------------------------------------------------------------
-// Witch ability
-// ---------------------------------------------------------------------------
-
-describe("SetNightTarget — Witch once-per-game restriction", () => {
-  const action = WEREWOLF_ACTIONS[WerewolfAction.SetNightTarget];
-
-  function makeWitchGame(witchAbilityUsed = false) {
-    const ts: WerewolfTurnState = {
-      turn: 2,
-      phase: {
-        type: WerewolfPhase.Nighttime,
-        startedAt: 1000,
-        nightPhaseOrder: [WerewolfRole.Witch],
-        currentPhaseIndex: 0,
-        nightActions: {},
-      },
-      deadPlayerIds: [],
-      ...(witchAbilityUsed ? { witchAbilityUsed: true } : {}),
-    };
-    return makePlayingGame(ts, {
-      players: [
-        { id: "p1", name: "Alice", sessionId: "s1", visibleRoles: [] },
-        { id: "p2", name: "Bob", sessionId: "s2", visibleRoles: [] },
-        { id: "p3", name: "Charlie", sessionId: "s3", visibleRoles: [] },
-      ],
-      roleAssignments: [
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Witch },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-      ],
-    });
-  }
-
-  it("Witch can set target when ability not yet used", () => {
-    const game = makeWitchGame(false);
-    expect(action.isValid(game, "p1", { targetPlayerId: "p2" })).toBe(true);
-  });
-
-  it("Witch cannot set target when ability already used", () => {
-    const game = makeWitchGame(true);
-    expect(action.isValid(game, "p1", { targetPlayerId: "p2" })).toBe(false);
-  });
-});
-
-describe("ConfirmNightTarget — sets witchAbilityUsed for Witch", () => {
-  const confirmAction = WEREWOLF_ACTIONS[WerewolfAction.ConfirmNightTarget];
-  const setTargetAction = WEREWOLF_ACTIONS[WerewolfAction.SetNightTarget];
-
-  it("sets witchAbilityUsed on the turn state after Witch confirms", () => {
-    const ts: WerewolfTurnState = {
-      turn: 2,
-      phase: {
-        type: WerewolfPhase.Nighttime,
-        startedAt: 1000,
-        nightPhaseOrder: [WerewolfRole.Witch],
-        currentPhaseIndex: 0,
-        nightActions: {},
-      },
-      deadPlayerIds: [],
-    };
-    const game = makePlayingGame(ts, {
-      players: [
-        { id: "p1", name: "Alice", sessionId: "s1", visibleRoles: [] },
-        { id: "p2", name: "Bob", sessionId: "s2", visibleRoles: [] },
-        { id: "p3", name: "Charlie", sessionId: "s3", visibleRoles: [] },
-      ],
-      roleAssignments: [
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Witch },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-      ],
-    });
-
-    setTargetAction.apply(game, { targetPlayerId: "p2" }, "p1");
-    expect(confirmAction.isValid(game, "p1", undefined)).toBe(true);
-    confirmAction.apply(game, null, "p1");
-
-    const resultTs = (game.status as { turnState: WerewolfTurnState })
-      .turnState;
-    expect(resultTs.witchAbilityUsed).toBe(true);
-  });
-
-  it("witchAbilityUsed is carried forward into the next night", () => {
-    const startNight = WEREWOLF_ACTIONS[WerewolfAction.StartNight];
-    const startDay = WEREWOLF_ACTIONS[WerewolfAction.StartDay];
-
-    const ts: WerewolfTurnState = {
-      turn: 2,
-      phase: {
-        type: WerewolfPhase.Nighttime,
-        startedAt: 1000,
-        nightPhaseOrder: [WerewolfRole.Witch],
-        currentPhaseIndex: 0,
-        nightActions: {
-          [WerewolfRole.Witch]: { targetPlayerId: "p2", confirmed: true },
-        },
-      },
-      deadPlayerIds: [],
-      witchAbilityUsed: true,
-    };
-    const game = makePlayingGame(ts, {
-      players: [
-        { id: "p1", name: "Alice", sessionId: "s1", visibleRoles: [] },
-        { id: "p2", name: "Bob", sessionId: "s2", visibleRoles: [] },
-        { id: "p3", name: "Charlie", sessionId: "s3", visibleRoles: [] },
-      ],
-      roleAssignments: [
-        { playerId: "p1", roleDefinitionId: WerewolfRole.Witch },
-        { playerId: "p2", roleDefinitionId: WerewolfRole.Villager },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Villager },
-      ],
-    });
-
-    startDay.apply(game, null, "owner-1");
-    startNight.apply(game, null, "owner-1");
-
-    const newTs = (game.status as { turnState: WerewolfTurnState }).turnState;
-    expect(newTs.witchAbilityUsed).toBe(true);
-  });
-});
-
-// ---------------------------------------------------------------------------
-// SetNightTarget / ConfirmNightTarget — suffixed repeat group phase (Wolf Cub bonus)
-// ---------------------------------------------------------------------------
-
-const WOLF_CUB_BONUS_PHASE_KEY = `${WerewolfRole.Werewolf as string}:2`;
-
-function makeWolfCubBonusGame(
-  nightActions: Record<string, AnyNightAction> = {},
-): Game {
-  return makeTeamGame({
-    turn: 2,
-    phase: {
-      type: WerewolfPhase.Nighttime,
-      startedAt: 1000,
-      nightPhaseOrder: [WerewolfRole.Werewolf, WOLF_CUB_BONUS_PHASE_KEY],
-      currentPhaseIndex: 1,
-      nightActions,
-    },
-    deadPlayerIds: [],
-  });
-}
-
-describe("SetNightTarget — suffixed repeat group phase", () => {
-  const action = WEREWOLF_ACTIONS[WerewolfAction.SetNightTarget];
-
-  it("rejects targeting the same player as the first phase's suggestedTargetId", () => {
-    const game = makeWolfCubBonusGame({
-      [WerewolfRole.Werewolf]: {
-        votes: [
-          { playerId: "w1", targetPlayerId: "p3" },
-          { playerId: "w2", targetPlayerId: "p3" },
-        ],
-        suggestedTargetId: "p3",
-        confirmed: true,
-      },
-    });
-    expect(action.isValid(game, "w1", { targetPlayerId: "p3" })).toBe(false);
-  });
-
-  it("allows targeting a different player in the second phase", () => {
-    const game = makeWolfCubBonusGame({
-      [WerewolfRole.Werewolf]: {
-        votes: [
-          { playerId: "w1", targetPlayerId: "p3" },
-          { playerId: "w2", targetPlayerId: "p3" },
-        ],
-        suggestedTargetId: "p3",
-        confirmed: true,
-      },
-    });
-    expect(action.isValid(game, "w1", { targetPlayerId: "p4" })).toBe(true);
-  });
-});
-
-describe("ConfirmNightTarget — suffixed repeat group phase", () => {
-  const action = WEREWOLF_ACTIONS[WerewolfAction.ConfirmNightTarget];
-
-  it("succeeds for the second Werewolf phase when all alive participants have voted", () => {
-    const game = makeWolfCubBonusGame({
-      [WerewolfRole.Werewolf]: {
-        votes: [
-          { playerId: "w1", targetPlayerId: "p3" },
-          { playerId: "w2", targetPlayerId: "p3" },
-        ],
-        suggestedTargetId: "p3",
-        confirmed: true,
-      },
-      [WOLF_CUB_BONUS_PHASE_KEY]: {
-        votes: [
-          { playerId: "w1", targetPlayerId: "p4" },
-          { playerId: "w2", targetPlayerId: "p4" },
-        ],
-        suggestedTargetId: "p4",
-      },
-    });
-    expect(action.isValid(game, "w1", null)).toBe(true);
-  });
-
-  it("second phase confirm is valid regardless of first phase confirmed state", () => {
-    const game = makeWolfCubBonusGame({
-      [WerewolfRole.Werewolf]: {
-        votes: [
-          { playerId: "w1", targetPlayerId: "p3" },
-          { playerId: "w2", targetPlayerId: "p3" },
-        ],
-        suggestedTargetId: "p3",
-        confirmed: true,
-      },
-      [WOLF_CUB_BONUS_PHASE_KEY]: {
-        votes: [
-          { playerId: "w1", targetPlayerId: "p4" },
-          { playerId: "w2", targetPlayerId: "p4" },
-        ],
-        suggestedTargetId: "p4",
-      },
-    });
-    // The active phase key is the suffixed key; confirmed on the base key must not block.
-    expect(action.isValid(game, "w1", null)).toBe(true);
-  });
-});
-
-// ---------------------------------------------------------------------------
-// StartNight — Wolf Cub bonus phase appears the turn after death, then is gone
-// ---------------------------------------------------------------------------
-
-describe("StartNight — Wolf Cub bonus phase lifecycle", () => {
-  const startNight = WEREWOLF_ACTIONS[WerewolfAction.StartNight];
-  const startDay = WEREWOLF_ACTIONS[WerewolfAction.StartDay];
-  const markDead = WEREWOLF_ACTIONS[WerewolfAction.MarkPlayerDead];
-
-  function makeWolfCubGame(): Game {
-    return {
-      id: "game-1",
-      lobbyId: "lobby-1",
-      gameMode: GameMode.Werewolf,
-      status: {
-        type: GameStatus.Playing,
-        turnState: {
-          turn: 1,
-          phase: {
-            type: WerewolfPhase.Daytime,
-            startedAt: 1000,
-            nightActions: {},
-          },
-          deadPlayerIds: [],
-        } satisfies WerewolfTurnState,
-      },
-      players: [
-        { id: "w1", name: "Wolf1", sessionId: "sw1", visibleRoles: [] },
-        { id: "c1", name: "Cub", sessionId: "sc1", visibleRoles: [] },
-        { id: "p3", name: "Seer", sessionId: "s3", visibleRoles: [] },
-      ],
-      roleAssignments: [
-        { playerId: "w1", roleDefinitionId: WerewolfRole.Werewolf },
-        { playerId: "c1", roleDefinitionId: WerewolfRole.WolfCub },
-        { playerId: "p3", roleDefinitionId: WerewolfRole.Seer },
-      ],
-      configuredRoleSlots: [],
-      showRolesInPlay: ShowRolesInPlay.None,
-      ownerPlayerId: "owner-1",
-    };
-  }
-
-  it("bonus phase appears in nightPhaseOrder the turn after Wolf Cub death", () => {
-    const game = makeWolfCubGame();
-    markDead.apply(game, { playerId: "c1" }, "owner-1");
-    startNight.apply(game, null, "owner-1");
-    const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-    const phase = ts.phase as WerewolfNighttimePhase;
-    expect(
-      phase.nightPhaseOrder.filter((k) =>
-        k.startsWith(WerewolfRole.Werewolf as string),
-      ),
-    ).toHaveLength(2);
-    expect(phase.nightPhaseOrder).toContain(WOLF_CUB_BONUS_PHASE_KEY);
-  });
-
-  it("bonus phase is NOT present on the subsequent night after it is consumed", () => {
-    const game = makeWolfCubGame();
-    // Turn 1 day: mark Wolf Cub dead, then transition to night (turn 2).
-    markDead.apply(game, { playerId: "c1" }, "owner-1");
-    startNight.apply(game, null, "owner-1");
-    // Turn 2 night → day: wolfCubDied flag is consumed by StartNight, not carried over.
-    startDay.apply(game, null, "owner-1");
-    // Turn 3 night: bonus phase should NOT appear.
-    startNight.apply(game, null, "owner-1");
-    const ts = (game.status as { turnState: WerewolfTurnState }).turnState;
-    const phase = ts.phase as WerewolfNighttimePhase;
-    expect(
-      phase.nightPhaseOrder.filter((k) =>
-        k.startsWith(WerewolfRole.Werewolf as string),
-      ),
-    ).toHaveLength(1);
-    expect(phase.nightPhaseOrder).not.toContain(WOLF_CUB_BONUS_PHASE_KEY);
-  });
-});
