+++ app/src/lib/game-modes/werewolf/actions/set-night-target.ts
@@ -1,0 +1,207 @@
+import type { Game, GameAction } from "@/lib/types";
+import { WerewolfPhase, isTeamNightAction, TargetCategory } from "../types";
+import type { TeamNightAction } from "../types";
+import {
+  currentTurnState,
+  validateActiveNightPlayer,
+  isGroupPhaseKey,
+  baseGroupPhaseKey,
+  getGroupPhasePlayerIds,
+  getGroupPhaseMemberIds,
+  computeSuggestedTarget,
+} from "../utils";
+import { WEREWOLF_ROLES, WerewolfRole } from "../roles";
+import type { WerewolfRoleDefinition } from "../roles";
+import { getPlayer } from "@/lib/player-utils";
+
+export const setNightTargetAction: GameAction = {
+  isValid(game: Game, callerId: string, payload: unknown) {
+    const ts = currentTurnState(game);
+    if (ts?.phase.type !== WerewolfPhase.Nighttime) return false;
+    if (ts.turn <= 1) return false;
+
+    const phase = ts.phase;
+    const { roleId: explicitPhaseKey, targetPlayerId } = payload as {
+      roleId?: unknown;
+      targetPlayerId?: unknown;
+    };
+
+    // Determine the phase key.
+    // Owner provides an explicit roleId/phaseKey; players use their own role.
+    let phaseKey: string;
+    const isOwner = callerId === game.ownerPlayerId;
+    if (isOwner) {
+      if (typeof explicitPhaseKey !== "string") return false;
+      if (!phase.nightPhaseOrder.includes(explicitPhaseKey)) return false;
+      phaseKey = explicitPhaseKey;
+    } else {
+      const result = validateActiveNightPlayer(game, callerId);
+      if (!result) return false;
+      phaseKey = result.activePhaseKey;
+
+      // Players cannot change once confirmed.
+      const existing = phase.nightActions[phaseKey];
+      if (existing?.confirmed) return false;
+    }
+
+    // Witch can only use her ability once per game.
+    if (
+      (phaseKey as WerewolfRole) === WerewolfRole.Witch &&
+      ts.witchAbilityUsed
+    )
+      return false;
+
+    // targetPlayerId undefined = clear; null = intentional skip; string = set target.
+    if (targetPlayerId === undefined) return true;
+    if (targetPlayerId === null) return true;
+    if (typeof targetPlayerId !== "string") return false;
+    if (targetPlayerId === game.ownerPlayerId) return false;
+    if (!game.players.some((p) => p.id === targetPlayerId)) return false;
+    if (ts.deadPlayerIds.includes(targetPlayerId)) return false;
+
+    // Suffixed repeat group phases (e.g. Wolf Cub bonus attack) cannot pick the
+    // same target as the base phase confirmed earlier this same night.
+    const baseKey = baseGroupPhaseKey(phaseKey);
+    if (baseKey !== phaseKey) {
+      const priorAction = phase.nightActions[baseKey];
+      if (
+        priorAction &&
+        isTeamNightAction(priorAction) &&
+        priorAction.suggestedTargetId === targetPlayerId
+      )
+        return false;
+    }
+
+    // Roles with preventRepeatTarget cannot target the same player twice in a row.
+    const phaseRoleDef = (
+      WEREWOLF_ROLES as Record<string, WerewolfRoleDefinition>
+    )[phaseKey];
+    if (
+      phaseRoleDef?.preventRepeatTarget &&
+      ts.lastTargets?.[phaseKey] === targetPlayerId
+    )
+      return false;
+
+    // Attack and Investigate roles cannot target themselves.
+    if (targetPlayerId === callerId) {
+      const callerAssignment = game.roleAssignments.find(
+        (a) => a.playerId === callerId,
+      );
+      const callerRoleDef = callerAssignment
+        ? (WEREWOLF_ROLES as Record<string, WerewolfRoleDefinition>)[
+            callerAssignment.roleDefinitionId
+          ]
+        : undefined;
+      if (
+        callerRoleDef?.targetCategory === TargetCategory.Attack ||
+        callerRoleDef?.targetCategory === TargetCategory.Investigate
+      )
+        return false;
+    }
+
+    // Group phases: cannot target players the caller knows are team members.
+    if (isGroupPhaseKey(phaseKey) && !isOwner) {
+      const caller = getPlayer(game.players, callerId);
+      const visibleTeammateIds = (caller?.visibleRoles ?? []).map(
+        (vr) => vr.playerId,
+      );
+      if (visibleTeammateIds.includes(targetPlayerId)) return false;
+    }
+    if (isGroupPhaseKey(phaseKey) && isOwner) {
+      const memberIds = getGroupPhaseMemberIds(game.roleAssignments, phaseKey);
+      if (memberIds.includes(targetPlayerId)) return false;
+    }
+
+    return true;
+  },
+  apply(game: Game, payload: unknown, callerId: string) {
+    const ts = currentTurnState(game);
+    if (ts?.phase.type !== WerewolfPhase.Nighttime) return;
+
+    const phase = ts.phase;
+    const { roleId: explicitPhaseKey, targetPlayerId } = payload as {
+      roleId?: string;
+      targetPlayerId?: string | null;
+    };
+
+    const phaseKey =
+      explicitPhaseKey ?? phase.nightPhaseOrder[phase.currentPhaseIndex];
+    if (!phaseKey) return;
+
+    if (isGroupPhaseKey(phaseKey)) {
+      // Group phase: upsert a player's vote.
+      const existing = phase.nightActions[phaseKey];
+      const groupAction: TeamNightAction =
+        existing && isTeamNightAction(existing)
+          ? { ...existing, votes: [...existing.votes] }
+          : { votes: [] };
+
+      const isOwner = callerId === game.ownerPlayerId;
+      if (isOwner) {
+        // Owner override: set all alive phase participants' votes.
+        const aliveParticipantIds = getGroupPhasePlayerIds(
+          game.roleAssignments,
+          phaseKey,
+          ts.deadPlayerIds,
+        );
+        if (targetPlayerId === undefined) {
+          groupAction.votes = [];
+        } else if (targetPlayerId === null) {
+          groupAction.votes = aliveParticipantIds.map((pid) => ({
+            playerId: pid,
+            skipped: true as const,
+          }));
+        } else {
+          groupAction.votes = aliveParticipantIds.map((pid) => ({
+            playerId: pid,
+            targetPlayerId,
+          }));
+        }
+      } else if (targetPlayerId === undefined) {
+        // Remove caller's vote.
+        groupAction.votes = groupAction.votes.filter(
+          (v) => v.playerId !== callerId,
+        );
+      } else {
+        // Upsert caller's vote (target or intentional skip).
+        const existingVote = groupAction.votes.find(
+          (v) => v.playerId === callerId,
+        );
+        if (targetPlayerId === null) {
+          if (existingVote) {
+            delete existingVote.targetPlayerId;
+            existingVote.skipped = true;
+          } else {
+            groupAction.votes.push({ playerId: callerId, skipped: true });
+          }
+        } else if (existingVote) {
+          existingVote.targetPlayerId = targetPlayerId;
+          delete existingVote.skipped;
+        } else {
+          groupAction.votes.push({ playerId: callerId, targetPlayerId });
+        }
+      }
+
+      const suggested = computeSuggestedTarget(groupAction.votes);
+      if (suggested !== undefined) {
+        groupAction.suggestedTargetId = suggested;
+      } else {
+        delete groupAction.suggestedTargetId;
+      }
+      phase.nightActions[phaseKey] = groupAction;
+    } else {
+      // Solo phase.
+      if (targetPlayerId === undefined) {
+        // Clear: remove the action entirely (back to undecided state).
+        phase.nightActions = Object.fromEntries(
+          Object.entries(phase.nightActions).filter(([k]) => k !== phaseKey),
+        );
+      } else if (targetPlayerId === null) {
+        // Intentional skip.
+        phase.nightActions[phaseKey] = { skipped: true };
+      } else {
+        phase.nightActions[phaseKey] = { targetPlayerId };
+      }
+    }
+  },
+};
