+++ app/src/lib/game-modes/werewolf/actions/types.ts
@@ -1,0 +1,10 @@
+export enum WerewolfAction {
+  StartNight = "start-night",
+  StartDay = "start-day",
+  SetNightPhase = "set-night-phase",
+  SetNightTarget = "set-night-target",
+  ConfirmNightTarget = "confirm-night-target",
+  RevealInvestigationResult = "reveal-investigation-result",
+  MarkPlayerDead = "mark-player-dead",
+  MarkPlayerAlive = "mark-player-alive",
+}
